+++ src/assets/TablerExtrasAsset.php
@@ -8,19 +8,15 @@ class TablerExtrasAsset extends BaseAssetBundle
 {
     public $sourcePath = '@npm/tabler--core/dist';
 
-    public $css = [
-        'css/tabler-flags.css',
-        'css/tabler-socials.css',
-        'css/tabler-payments.css',
-        'css/tabler-marketing.css',
-        'css/tabler-themes.css',
-    ];
+    public $css = [];
 
-    public $js = [
-        'js/tabler-theme.min.js',
-    ];
+    public $js = [];
 
     public $depends = [
-        'bestyii\tabler\assets\TablerAsset',
+        'bestyii\tabler\assets\TablerFlagsAsset',
+        'bestyii\tabler\assets\TablerPaymentsAsset',
+        'bestyii\tabler\assets\TablerSocialsAsset',
+        'bestyii\tabler\assets\TablerMarketingAsset',
+        'bestyii\tabler\assets\TablerThemeAsset',
     ];
 }